# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Teranga GESCRIM API",
        "description": "API REST pour le système de Gestion de la Criminalité et de la Délinquance — Direction de la Sécurité Publique (DSP) du Sénégal.",
        "contact": {
            "name": "NAMASTECH",
            "email": "admin@gescrim.sn"
        },
        "license": {
            "name": "Propriétaire",
            "url": "https://gescrim.sn"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur de développement"
        }
    ],
    "paths": {
...
        "/api/dashboard/stats": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "KPIs globaux du tableau de bord",
                "operationId": "654e1617f74d2b9e0654ec77620d960b",
                "parameters": [
                    {
                        "name": "annee",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "mois",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 12,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "region_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "departement_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "commune_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "service_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Statistiques retournées"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/dashboard/infractions-par-region": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Infractions par région",
                "operationId": "4372a6642330b2f060b8d4968edbb882",
                "parameters": [
                    {
                        "name": "annee",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "mois",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Infractions groupées par région"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/dashboard/accidents-par-type": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Accidents groupés par type",
                "operationId": "2d599e8bbc5b8a3f97f111fcda065ae7",
                "parameters": [
                    {
                        "name": "annee",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "mois",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Accidents par type"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/dashboard/tendances-mensuelles": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Tendances mensuelles infractions et accidents",
                "operationId": "44c64542dfee6fa89e077cda86f006c9",
                "parameters": [
                    {
                        "name": "annee",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "mois",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Données mensuelles pour graphiques"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Entrez votre token JWT (sans le mot Bearer)"
            }
        }
    },
...
}